move global style into theme folder ui comp folder is for low level components only
